requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
pillow>=9.0.0
opencv-python>=4.5.3
mediapipe>=0.8.9
rembg>=2.0.0
python-dotenv>=0.19.0
redis>=4.0.0
pydantic>=1.8.2 